non-linear: disable linear path chaining
The preprocessing step of chaining linear paths is problematic in
incremental mode, because we remove the chained rules/facts, which
cuts-off paths that might be reachable via a non-linear rule. For
example, consider the CHC problem:

    fib(1,1)
    fib(2,1)
    fib(a,c) /\ fib(b,d) /\ b=a+1 ==> f(b+1,c+d)
    fib(5,5) ==> false

The fact `fib(5,5)` is reachable so the problem should be UNSAT.
However, when we compute the resolvents of the non-linear clause with
all facts we get:

    fib(0,c) ==> fib(2,c+1)              (R1)
    fib(1,c) ==> fib(3,c+1)              (R2)
    fib(2,d) ==> fib(3,d+1)              (R3)
    fib(3,y) ==> fib(4,d+1)              (R4)

If we now chain linear paths we get:

    [fib(1,1), R2] = fib(3,2)
    [fib(2,1), R3] = fib(3,2)

    [fib(3,2), R4] = fib(4,3)

At this point we remove all facts except `fib(4,3)` because in the
linear context we can assume that all paths from these facts lead
to `fib(4,3)` anyway. However in the non-linear context we need to
keep `fib(3,2)` to derive `fib(5,5)`.

After disabling linear path chaining we can prove UNSAT of 13 new
problem instances in the LIA benchmark that previously gave UNKNOWN.
However, there is also a regression. For the following instances we
now get a timeout although we previously got UNSAT: 137,141,336

QUESTION:
We could add the chained rules to the ITS without removing
the rules that are part of the chain. However, that's potentially a
lot of rules and is this really cheaper re-generating these rules
during search? Maybe we can just add the maximal paths instead of
every prefix of every path.