propagate program variable equalities
Previously, when simplifying boolean expressions, only temporary
boolean variables were propagated. For example in

  (b1 \/ b2) /\ !b2 ==> F(b1)

the temporary variable `b2` can be propagated, which simplifies the
clause to

  b1 ==> F(b1)

But program variables were not propagated. For example

  (b1 \/ b2) /\ !b2 ==> F(b1, b2)

would not propagate `b2` since it's a program variable.
When propagating program variables we have to make sure to not
eliminate them completely like temporary variables. Otherwise we
loose information. Instead the expected result is:

  b1 /\ !b2 ==> F(b1, b2)

This simplification can speed up ADCL because potentially fewer
conjunctive variants of the clause have to generated
(observed in LIA instance 69).