non-linear: introduce "constraint tiers"
Introduce constraint tiers: Linear, Polynomial, Exponential. Here "linear" does
NOT refer the plurality of left-hand-side predicates but the clause constraint.
For example:

 - linear constraint      : F(x) /\ 2*x=4 ==> F(x)
 - polynomial constraint  : F(x) /\ x^2=4 ==> F(x)
 - exponential constraint : F(x) /\ 2^x=4 ==> F(x)

The distinction is for optimization. Because the constraint tiers are successively
harder to deal with for SMT solvers, we can first try to solve a CHC problem by only
considering "easy" linear constraints first and only consider harder polynomial and
exponential constraints if necessary.