move syntactic redundancy check into linear solver
The commit titled:

    non-linear: detect redundant (up to renaming) facts

introduced a syntactic redundancy check in the non-linear solver.
This allows the non-linear solver to filter out syntactially
redundant facts derived by the linear solver. Thus, we do fewer
resolution steps in the non-linear solver and produce fewer rules
that are passed to the linear solver.

However, the linear solver still spends a lot of time deriving
these redundant facts. By moving the redundancy check into the
linear solver, it can backtrack immediatly when encountering a
redundant fact.

With that we get `unsat` on 7 new instances.