non-linear: eliminate unilaterally resolvable CHCs
If a predicate only occurs on the RHS of a single CHC, then there is
only one choice how to eliminate the predicate on the LHS of other CHCs.
Thus, we can pre-compute all resolvents where one of parent clauses has
a unique RHS.

Previously the CHC parser would directly output an ITS instance.
But it's hard to reason about the side effect of deleting/modifying rules
in the ITS. It's more convenient to preprocess the CHCs first and then
construct the ITS instance from it. Thus, the CHC parser now outputs
a set of clauses.

With this change we get:

 - chc comp 2022: +16 UNSAT
 - chc comp 2023: +9  UNSAT