fix: still redundant resolvents in non-linear solver
Assume we have facts F,G,H and a non-linear CHC:

  F /\ G /\ H ==> false

When we recursively compute all resolvents, the iterator on the LHS
predicates should match the call tree depth. So at depth 1 the iterator
should point to the first predicate. At depth 2 the iterator should always
point on the second predicate. Let `*` denote the iterator position and
let `_` denote an eliminated predicate, then the (partial) call tree should
look like this:

  F* /\ G /\ H ==> false

  -->  _ /\ G* /\ H ==> false

      --> _ /\ _ /\ H* ==> false

      (...)

  --> F /\ G* /\ H ==> false

      --> F /\ _ /\ H* ==> false

          --> F /\ _ /\ _ ==> false

      (...)

However, when we compute a resolvent the iterator was "reset" to the first
predicate, which is a problem when we are in a sub-tree, where the iterator
is not pointing on the first non-eliminated predicate. That is, we actually
got:

  F* /\ G /\ H ==> false

  -->  _ /\ G* /\ H ==> false

      --> _ /\ _ /\ H* ==> false                   (original resolvent)

      (...)

  --> F /\ G* /\ H ==> false

      --> F* /\ _ /\ H ==> false                   (iterator reset)

          --> _ /\ _ /\ H* ==> false               (redundant resolvent)

      (...)

The iterator is "reset" because iterators are only defined for a concrete
collection. The resolvent stores its LHS predicate in a new collection.
We fix this by using an normal index on the LHS predicates instead of an
iterator.

Another way we derive redundant resolvents is when clauses are "symmetric".
For example, exchanging `F(x)` and `F(y)` in

  F(x) /\ F(y) /\ F(z) ==> G(x+y)

does not change the meaning of the clause. Now, with the single fact `F(1)`
we would get the derivation:

  F(x)* /\ F(y) /\ F(z) ==> G(x+y)

  --> ___ /\ F(y) /\ F(z) ==> G(1+y)

      (...)

  --> F(x) /\ F(y)* /\ F(z) ==> G(x+y)

    --> F(x) /\ ___ /\ F(z) ==> G(x+1)

    (...)

We can detect that the two resolvents

  F(y) /\ F(z) ==> G(1+y)
  F(x) /\ F(z) ==> G(x+1)

are syntactially equivalent (up-to renmaing). And in particular, all resolvents
that derived from these clauses are also equivalent. So we now prune the call tree
when we detect a redundant resolvent.