non-linear: merge facts with same RHS predicate
Computing all resolvents is exponential in the number of LHS predicates
of a non-linear clause. So it's worth eliminating as many predicates as
possible. In preprocessing we already eliminate predicates if there is
only one resolution candidate.

By merging facts we can reduce resolution candidates. For example, here
have two candidates for `F(x)`:

    x=0 ==> F(x)

    x>0 ==> F(x)

we can merge these facts into one clause by disjoining the constraints:

    (x=0 \/ x>0) ==> F(x)

This can make all facts unilaterally resolvable as long as the RHS
predicate does not also occur on the RHS of a rule.

Arguably, that just off-loads dealing with disjunctions to the linear
solver. We also loose some CHC structure pushing disjunctions into the
clause constraint. However, ADCL is designed to deal with disjunctions.
And computing all resolvents in the non-linear solver is worst-case
exponential in the number of LHS predicates, so it's probably worth it.

Benchmark results:

 - CHC comp 2022: +6 UNSAT
 - CHC comp 2023: +3 UNSAT