GitHub action for running benchmarks
Since running all benchmarks takes so long, it's easy to miss regressions
or progressions due to changes. Then later when they are spotted, it can
be time consuming to figure out which change was responsible. By setting
up a GitHub action, that runs on every pushed commit we get full benchmark
results that are always associated with a change.

Running the benchmarks requires building the LoAT binary, and that in turn
requires building all the dependencies (z3, yices, etc.) which takes very
long and is very fragile. Presumably, the dependencies don't change that
often though and don't actually depend on the sources in the LoAT repo.
Thus, it makes sense to prepare a base docker image with all dependencies
that is only re-build when needed. The setup looks as follows:

 * the Dockerfile in `docker/loat-base-image/` describes all dependencies
   but does not build LoAT itself.
 * a dedicated GitHub action (`.github/workflows/build-base-image.yml`)
   that automatically runs iff changes in `docker/loat-base-image/`
   are pushed to `master` (can also be triggered manually)
 * the action builds the image and pushes it to Docker Hub.

With the base image on Docker Hub, the benchmark action can just pull it
and build the LoAT binary inside it. The base images can also be used
locally, so it's not/less necessary to wait for these slow builds to
complete locally. The base images are also tagged with a commit SHA. So
older base images are still accessible. That's particularly useful when
working on a branch where the dependencies older/newer compared to `master`.
It's very annoying when all the dependencies have to be re-build just
when switching between `master` and feature branch.

To describe all dependendencies together I put everything back into a
single Dockerfile (sorry).

The benchmark action runs on every push to every branch including pull
requests from external forks. At the moment it runs on the LIA-Lin
benchmark for CHC comp 22 and 23. As a nice side effect is that the built
LoAT binary of each run can be downloaded through the GitHub UI.